requirements.txt
asgiref==3.8.1
Django==3.2.21
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.6.1
pymongo==4.8.0
pytz==2024.1
sqlparse==0.2.4
typing-extensions>=4
